add TypeNodeAnalyzer
