+++ packages/coding-standard/config/symplify-rules.neon
@@ -247,3 +247,4 @@ services:
     - Symplify\CodingStandard\PHPStan\ParentGuard\ParentMethodReturnTypeResolver
     - Symplify\CodingStandard\PHPStan\NodeComparator
     - Symplify\CodingStandard\NodeAnalyzer\DependencyNodeAnalyzer
+    - Symplify\CodingStandard\NodeAnalyzer\TypeNodeAnalyzer
